feat: update datafusion to 49

+++ src/config/src/utils/async_file.rs
@@ -40,16 +40,15 @@ pub async fn is_exists(path: impl AsRef<Path>) -> bool {
 #[inline(always)]
 pub async fn get_file_contents(
     path: impl AsRef<Path>,
-    range: Option<Range<usize>>,
+    range: Option<Range<u64>>,
 ) -> Result<Vec<u8>, std::io::Error> {
     let mut file = File::open(path).await?;
     let data = if let Some(range) = range {
         let to_read = range.end - range.start;
-        let mut buf = Vec::with_capacity(to_read);
-        file.seek(std::io::SeekFrom::Start(range.start as u64))
-            .await?;
-        let read = file.take(to_read as u64).read_to_end(&mut buf).await?;
-        if read != to_read {
+        let mut buf = Vec::with_capacity(to_read as usize);
+        file.seek(std::io::SeekFrom::Start(range.start)).await?;
+        let read = file.take(to_read).read_to_end(&mut buf).await?;
+        if read != to_read as usize {
             return Err(std::io::Error::new(
                 std::io::ErrorKind::UnexpectedEof,
                 format!("Expected to read {to_read} bytes, but read {read} bytes"),
